python/requirements.txt
openai-whisper>=20240930
pyannote.audio>=3.1.0
resemblyzer>=0.1.3
numpy>=1.26.0
soundfile>=0.12.1
torch>=2.0.0
torchaudio>=2.0.0
